requirements/base.txt
django==4.2.5
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-environ==0.11.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
django_celery_beat==2.5.0
flower==2.0.1
requests==2.31.0
djoser==2.2.2
Pillow==9.4.0

